requirements.txt
absl-py==2.0.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.1
cchardet==2.1.7
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
fastprogress==1.0.3
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
jieba==0.42.1
joblib==1.3.2
keras==2.14.0
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
keras-self-attention==0.51.0
keras-transformer==0.40.0
kiwisolver==1.4.5
ktrain==0.38.0
langdetect==1.0.9
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==1.7.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
sqlparse==0.4.4
syntok==1.4.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.3.0
threadpoolctl==3.2.0
tika==2.6.0
tokenizers==0.14.1
tqdm==4.66.1
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.0
Whoosh==2.7.4
wrapt==1.14.1
zipp==3.17.0
